improve fetched results functionality
