+++ src/components/waves/organisms/Tabs/Generic/index.jsx
@@ -4,6 +4,12 @@ import React from 'react'
 // Components
 //import CreateTab from '../../organisms/Modals/Create/tab'
 
+// Fade In Animation
+import FadeIn from 'react-fade-in'
+
+// Redux
+import { connect } from 'react-redux'
+
 // Forms
 import DefFields from './defineFields'
 
@@ -20,7 +26,22 @@ class TabGeneric extends React.Component{
         // Check if fields are set
         if(true){
             return(
-                <DefFields title={tab.title} />
+                <div>
+                    {tab.fields.length > 0 &&
+                            <div>
+                                {tab.fields.map((field, index) => {
+                                    return(
+                                        <FadeIn key={index}>
+                                            <div>
+                                                {field.field}
+                                            </div>
+                                        </FadeIn>
+                                    );
+                                })}
+                            </div>
+                        }
+                    <DefFields title={tab.title} />
+                </div>
             )
         } else {
             return null;
@@ -28,4 +49,10 @@ class TabGeneric extends React.Component{
     }
 }
 
-export default TabGeneric;
+const mapStateToProps = (state) => {
+    return {
+        profile: state.firebase.profile,
+    }
+}
+
+export default connect(mapStateToProps)(TabGeneric);
